int로 타입변경
현재를 long 타입이 필요가 없는것 같아 다시 int형으로 변경했습니다.

+++ src/main/java/com/codessquad/qna/controller/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
     }
 
     @GetMapping("{id}/form")
-    public String viewUpdateUserForm(@PathVariable long id, Model model){
+    public String viewUpdateUserForm(@PathVariable int id, Model model){
         for(User user : users){
             if(user.getId()==id) {
                 model.addAttribute("user", user);
@@ -53,8 +53,8 @@ public class UserController {
     }
 
     @PostMapping("{{id}}/update")
-    public String updateUser(@PathVariable long id, Model model){
-        model.addAttribute("user",users.get((int)id-1));
+    public String updateUser(@PathVariable int id, Model model){
+        model.addAttribute("user",users.get(id-1));
         return "redirect:/users";
     }
 }
